Refactor - extract SaveProviderUserService
