+++ app/forms/provider_interface/provider_user_invitation_wizard.rb
@@ -84,7 +84,7 @@ module ProviderInterface
     end
 
     def save!
-      SaveNewProviderUserService.new(self).call
+      SaveProviderUserService.new(self).call
     end
 
     def save_state!
@@ -127,80 +127,6 @@ module ProviderInterface
 
     def any_provider_needs_permissions_setup
       next_provider_needing_permissions_setup.present?
-    end
-  end
-
-  class SaveNewProviderUserService
-    attr_accessor :wizard
-
-    def initialize(wizard)
-      self.wizard = wizard
-    end
-
-    def call
-      if email_exists?
-        update_user
-      else
-        create_user
-      end
-    end
-
-  private
-
-    def email_exists?
-      ProviderUser.find_by(email_address: wizard.email_address).present?
-    end
-
-    def update_user
-      existing_user = ProviderUser.find_by(email_address: wizard.email_address)
-      existing_user.update(
-        email_address: wizard.email_address,
-        first_name: wizard.first_name,
-        last_name: wizard.last_name,
-      )
-      update_provider_permissions(existing_user)
-    end
-
-    def create_user
-      user = ProviderUser.create(
-        email_address: wizard.email_address,
-        first_name: wizard.first_name,
-        last_name: wizard.last_name,
-      )
-      create_provider_permissions(user)
-      user
-    end
-
-    def create_provider_permissions(user)
-      wizard.provider_permissions.each do |provider_id, permission|
-        provider_permission = ProviderPermissions.new(
-          provider_id: provider_id,
-          provider_user_id: user.id,
-        )
-        permission['permissions'].each do |permission_name|
-          provider_permission.send("#{permission_name}=".to_sym, true)
-        end
-        provider_permission.save!
-      end
-    end
-
-    def update_provider_permissions(user)
-      wizard.provider_permissions.each do |provider_id, permission|
-        provider_permission = ProviderPermissions.find_or_initialize_by(
-          provider_id: provider_id,
-          provider_user_id: user.id,
-        )
-        %w[manage_users make_decisions].each do |permission_type|
-          provider_permission.send(
-            "#{permission_type}=",
-            permission['permissions'].include?(permission_type),
-          )
-        end
-        provider_permission.save!
-      end
-
-      deleted_permissions = user.provider_permissions.select { |permission| wizard.provider_permissions[permission.provider_id.to_s].blank? }
-      deleted_permissions.each(&:destroy)
     end
   end
 end
